fix: cast node last seen age to integer in ProcessCommandTimeouts
- Updated the calculation of node last seen age in the ProcessCommandTimeouts command to explicitly cast the result to an integer, ensuring consistent data type handling.

+++ backend/laravel/app/Console/Commands/ProcessCommandTimeouts.php
@@ -86,7 +86,7 @@ class ProcessCommandTimeouts extends Command
                     if ($command->zone_id) {
                         $nodeLastSeenAt = $command->node_last_seen_at ? Carbon::parse($command->node_last_seen_at) : null;
                         $nodeLastSeenAgeSec = $nodeLastSeenAt instanceof CarbonInterface
-                            ? max(0, $nodeLastSeenAt->diffInSeconds(now()))
+                            ? (int) max(0, $nodeLastSeenAt->diffInSeconds(now()))
                             : null;
                         $nodeStatus = $command->node_status ? strtolower((string) $command->node_status) : null;
